Move CollectorFactory from core to sqlobject
