+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/collector/OptionalBuilder.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core.collector;
+package org.jdbi.v3.sqlobject.collector;
 
 import java.util.Optional;
